web/django/requirements.txt
Django==3.0.3
gunicorn==20.0.4
psycopg2-binary==2.8.4
djangorestframework==3.11.0
markdown==3.2.1
django-filter==2.2.0
djoser==2.0.3
djangorestframework_simplejwt==4.4.0
django-simple-history==2.8.0
